Add an agent injecting mutating webhook to the traffic manager
This commit adds a mutating webhook that injects a traffic-agent to
pods that have the following annotation:

    telepresence.getambassador.io/traffic-agent: enabled

The installer will henceforth ignore deployments that have this
annotation and consider them to already have an agent installed.

The webhook is using `kates.List()` to list Service objects which
means that the cluster must be configured to allow the traffic-manager
to do that. A sample `k8s/agent-injector-rbac.yaml` is included that
provides the needed credentials.

vLuke:
 - Split in to separate commits
 - Move filling of the default agent name to be in envconfig.go instead of agent_injector.go
 - Tidy the naming and ordering of envconfig fields
 - Fix linter complaint regarding import order
 - Say "corev1" instead of "v1"
 - Minimize the diff in pkg/install; this means dropping some naming,
   formatting, and order changes.
 - Arrange AgentContainer's arguments 1-per-line; those lines were very
   long.
 - Remove mis-copy-pasted comment from agent-injector-rbac.yaml that was
   originally destined to be removed in a later commit.
 - Add a changelog entry (instead of putting it off to a later commit)

Signed-off-by: Thomas Hallgren <thomas@datawire.io>
Signed-off-by: Luke Shumaker <lukeshu@datawire.io>