connector: Move some of the 'install' code in to separate files
This is in preparation to move them to a separate package so that they
can be shared with the manager.  This commit makes essentially no
modifications, and is a simple cut/paste job.  The one exception to
that is when pulling part of a function out to be its own function.

vLuke:
 - Split in to separate commit

Signed-off-by: Thomas Hallgren <thomas@datawire.io>
Signed-off-by: Luke Shumaker <lukeshu@datawire.io>