Make creation and deletion of manager resources automatic
With the introduction of the mutating webhook for agent injection, a
slew of new resources were introduced. We now have the following nine
associated with the traffic-manager:

- Namespace
- ServiceAccount
- ClusterRole
- ClusterRoleBinding
- Secret (for the webhook TLS)
- Deployment
- Service (headless for the traffic manager)
- Service (proxy for the agent-injector)
- MutatingWebhookConfiguration

This commit invents a `resource.Instance` as a common pattern for
handling the CRUD and logging of those resources in a unified way
so that it's ensured that the resources are created in the given
order and deleted in the reverse order (we still retain the namespace).

vLuke:
 - Mention config defaults changing in the changelog

Signed-off-by: Thomas Hallgren <thomas@datawire.io>
Signed-off-by: Luke Shumaker <lukeshu@datawire.io>